app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
Django==3.2.6
dnspython==2.2.1
Faker==15.3.4
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
mongoengine==0.24.2
platformdirs==2.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pylint==2.15.8
pylint-mongoengine==0.4.0
pylint-plugin-utils==0.7
pymongo==3.12.3
gunicorn==20.1.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
djongo==1.3.6
djangorestframework==3.14.0
django-crispy-forms==1.14.0